Migrate geocoding + logging
- add test logs
- migrate geocoding from Service to MapboxTelemetry
- tweaks to system
